32、第一遍测试并修改bug

+++ app/Models/AdminUser.php
@@ -23,6 +23,11 @@ class AdminUser extends Administrator
 
         static::deleted(function ($model) {
 
+            // 删除教师、课程绑定关系
+            DB::table('teachers_courses')
+                ->where('teacher_id', $model->id)
+                ->delete();
+
             // 删除学生对应的分数
             DB::table('fraction')
                 ->where('student_id',  $model->id)
@@ -35,6 +40,46 @@ class AdminUser extends Administrator
 
             // 删除对应前台用户
             User::where('email', $model->username)->delete();
+
+            // 删除该用户发布的文章
+            $articles = DB::table('article')
+                ->where('user_id', $model->id)
+                ->get();
+
+            // 删除该文章对应的所有评论
+            if ($articles->count() > 0) {
+                foreach ($articles as $key => $value) {
+                    DB::table('comments')
+                        ->where('article_id', $value->id)
+                        ->delete();
+
+                    DB::table('article')
+                        ->where('id', $value->id)
+                        ->delete();
+                }
+            }
+
+            // 删除该用户发布的评论，
+            $comments = DB::table('comments')
+                ->where('user_id', $model->id)
+                ->get();
+
+            if ($comments->count() > 0) {
+                foreach ($comments as $key => $value) {
+                    // 如果是父级评论，连带着子级评论一起删除
+                    if ($value->level == 1) {
+                        // 删除子节点
+                        DB::table('comments')
+                            ->where('pid', $value->id)
+                            ->delete();
+                    }
+                    // 删除自身
+                    DB::table('comments')
+                        ->where('id', $value->id)
+                        ->delete();
+                }
+            }
+
 
         });
     }
